+++ src/language/ExpressionBuilder.ts
@@ -2,11 +2,14 @@ import {
   ExpressionFactory,
   ParameterType,
   ParameterTypeRegistry,
+  RegExps,
+  StringOrRegExp,
 } from '@cucumber/cucumber-expressions'
+import { LocationLink } from 'vscode-languageserver-types'
 
 import { createLocationLink, makeParameterType, sortLinks, syntaxNode } from './helpers.js'
 import { getLanguage } from './languages.js'
-import { SourceAnalyzer } from './SourceAnalyzer.js'
+import { SourceAnalyzer, SourceMatch } from './SourceAnalyzer.js'
 import {
   ExpressionBuilderResult,
   ExpressionLink,
@@ -49,8 +52,8 @@ export class ExpressionBuilder {
     )
 
     let parameterTypeLinks: ParameterTypeLink[] = []
-    for (const [language, matches] of parameterTypeMatches.entries()) {
-      const links = language.buildParameterTypeLinks(matches)
+    for (const [_, matches] of parameterTypeMatches.entries()) {
+      const links = buildParameterTypeLinksFromMatches(matches)
       parameterTypeLinks = parameterTypeLinks.concat(links)
       for (const { parameterType } of links) {
         defineParameterType(parameterType)
@@ -86,4 +89,46 @@ export class ExpressionBuilder {
       registry,
     }
   }
+}
+
+function buildParameterTypeLinksFromMatches(
+  parameterTypeMatches: readonly SourceMatch[]
+): readonly ParameterTypeLink[] {
+  const parameterTypeLinks: ParameterTypeLink[] = []
+  const propsByName: Record<string, ParameterTypeLinkProps> = {}
+  for (const { source, match } of parameterTypeMatches) {
+    const nameNode = syntaxNode(match, 'name')
+    const rootNode = syntaxNode(match, 'root')
+    const expressionNode = syntaxNode(match, 'expression')
+    if (nameNode && rootNode) {
+      const language = getLanguage(source.languageName)
+
+      const parameterTypeName = language.toParameterTypeName(nameNode)
+      const regExps = language.toParameterTypeRegExps(expressionNode)
+      const selectionNode = expressionNode || nameNode
+      const locationLink = createLocationLink(rootNode, selectionNode, source.uri)
+      const props: ParameterTypeLinkProps = (propsByName[parameterTypeName] = propsByName[
+        parameterTypeName
+      ] || { locationLink, regexpsList: [] })
+      props.regexpsList.push(regExps)
+    }
+  }
+  for (const [name, { regexpsList, locationLink }] of Object.entries(propsByName)) {
+    const regexps: StringOrRegExp[] = regexpsList.reduce<StringOrRegExp[]>((prev, current) => {
+      if (Array.isArray(current)) {
+        return prev.concat(...current)
+      } else {
+        return prev.concat(current)
+      }
+    }, [])
+    const parameterType = makeParameterType(name, regexps)
+    parameterTypeLinks.push({ parameterType, locationLink })
+  }
+
+  return parameterTypeLinks
+}
+
+type ParameterTypeLinkProps = {
+  regexpsList: RegExps[]
+  locationLink: LocationLink
 }
