+++ src/language/helpers.ts
@@ -59,48 +59,6 @@ export function createLocationLink(
   return locationLink
 }
 
-export function buildParameterTypeLinksFromMatches(
-  parameterTypeMatches: readonly SourceMatch[]
-): readonly ParameterTypeLink[] {
-  const parameterTypeLinks: ParameterTypeLink[] = []
-  const propsByName: Record<string, ParameterTypeLinkProps> = {}
-  for (const { source, match } of parameterTypeMatches) {
-    const nameNode = syntaxNode(match, 'name')
-    const rootNode = syntaxNode(match, 'root')
-    const expressionNode = syntaxNode(match, 'expression')
-    if (nameNode && rootNode) {
-      const language = getLanguage(source.languageName)
-
-      const parameterTypeName = language.toParameterTypeName(nameNode)
-      const regExps = language.toParameterTypeRegExps(expressionNode)
-      const selectionNode = expressionNode || nameNode
-      const locationLink = createLocationLink(rootNode, selectionNode, source.uri)
-      const props: ParameterTypeLinkProps = (propsByName[parameterTypeName] = propsByName[
-        parameterTypeName
-      ] || { locationLink, regexpsList: [] })
-      props.regexpsList.push(regExps)
-    }
-  }
-  for (const [name, { regexpsList, locationLink }] of Object.entries(propsByName)) {
-    const regexps: StringOrRegExp[] = regexpsList.reduce<StringOrRegExp[]>((prev, current) => {
-      if (Array.isArray(current)) {
-        return prev.concat(...current)
-      } else {
-        return prev.concat(current)
-      }
-    }, [])
-    const parameterType = makeParameterType(name, regexps)
-    parameterTypeLinks.push({ parameterType, locationLink })
-  }
-
-  return parameterTypeLinks
-}
-
-type ParameterTypeLinkProps = {
-  regexpsList: RegExps[]
-  locationLink: LocationLink
-}
-
 export function childrenToString(node: TreeSitterSyntaxNode, stringNodes: NodePredicate) {
   return node.children
     .filter(stringNodes)
